requirements.txt
DBUtils==1.3
PyMySQL==0.9.3
pyserial==3.4
